Refactor : remove almost all templating on dataType
