+++ core/base/bottleneckDistance/BottleneckDistance.h
@@ -3,11 +3,6 @@
 /// \author Maxime Soler <soler.maxime@total.com>
 #pragma once
 
-#ifndef diagramTuple
-#define diagramTuple                                                        \
-  std::tuple<int, ttk::CriticalType, int, ttk::CriticalType, dataType, int, \
-             dataType, float, float, float, dataType, float, float, float>
-#endif
 
 #ifndef matchingTuple
 #define matchingTuple std::tuple<int, int, double>
@@ -31,6 +26,9 @@
 #include <string>
 #include <tuple>
 
+typedef double dataType;
+typedef  std::tuple<int, ttk::CriticalType, int, ttk::CriticalType, dataType, int, dataType, float, float, float, dataType, float, float, float> diagramTuple;
+
 namespace ttk {
 
   using trackingTuple = std::tuple<int, int, std::vector<int>>;
@@ -46,7 +44,6 @@ namespace ttk {
 
     ~BottleneckDistance(){};
 
-    template <typename dataType>
     int execute(bool usePersistenceMetric);
 
     inline int setPersistencePercentThreshold(double t) {
@@ -142,26 +139,22 @@ namespace ttk {
     double ps_;
 
   private:
-    template <typename dataType>
     int computeBottleneck(const std::vector<diagramTuple> &d1,
                           const std::vector<diagramTuple> &d2,
                           std::vector<matchingTuple> &matchings,
                           bool usePersistenceMetric);
 
-    template <typename dataType>
     double computeGeometricalRange(const std::vector<diagramTuple> &CTDiagram1,
                                    const std::vector<diagramTuple> &CTDiagram2,
                                    int d1Size,
                                    int d2Size) const;
 
-    template <typename dataType>
     double computeMinimumRelevantPersistence(
       const std::vector<diagramTuple> &CTDiagram1,
       const std::vector<diagramTuple> &CTDiagram2,
       int d1Size,
       int d2Size) const;
 
-    template <typename dataType>
     void computeMinMaxSaddleNumberAndMapping(
       const std::vector<diagramTuple> &CTDiagram,
       int dSize,
@@ -173,7 +166,6 @@ namespace ttk {
       std::vector<int> &sadMap,
       dataType zeroThresh);
 
-    template <typename dataType>
     void buildCostMatrices(
       const std::vector<diagramTuple> &CTDiagram1,
       const std::vector<diagramTuple> &CTDiagram2,
@@ -191,14 +183,12 @@ namespace ttk {
       bool reverseSad,
       int wasserstein);
 
-    template <typename dataType>
     void solvePWasserstein(int nbRow,
                            int nbCol,
                            std::vector<std::vector<dataType>> &matrix,
                            std::vector<matchingTuple> &matchings,
                            Munkres &solver);
 
-    template <typename dataType>
     void solveInfinityWasserstein(int nbRow,
                                   int nbCol,
                                   int nbRowToCut,
@@ -207,7 +197,6 @@ namespace ttk {
                                   std::vector<matchingTuple> &matchings,
                                   GabowTarjan &solver);
 
-    template <typename dataType>
     dataType buildMappings(const std::vector<matchingTuple> &inputMatchings,
                            bool transposeGlobal,
                            bool transposeLocal,
@@ -216,9 +205,5 @@ namespace ttk {
                            const std::vector<int> &m2,
                            int wasserstein);
   };
-
-// Include in namespace ttk
-#include <BottleneckDistanceImpl.h>
-#include <BottleneckDistanceMainImpl.h>
 
 } // namespace ttk
