+++ core/base/bottleneckDistance/GabowTarjanImpl.h
@@ -3,8 +3,10 @@
 #include "MatchingGraph.h"
 #include <iostream>
 #include <vector>
+#include "GabowTarjan.h"
 
-template <typename dataType>
+namespace ttk {
+
 bool GabowTarjan::DFS(int v) {
   if(v < 0)
     return true;
@@ -13,7 +15,7 @@ bool GabowTarjan::DFS(int v) {
   for(unsigned int i = 0; i < Connections[v].size(); ++i) {
     int u = Connections[v][i];
     if(Layers[Pair[u] + 1] == Layers[v + 1] + 1) {
-      if(DFS<dataType>(Pair[u])) {
+      if(DFS(Pair[u])) {
         Pair[u] = v;
         Pair[v] = u;
         return true;
@@ -26,7 +28,6 @@ bool GabowTarjan::DFS(int v) {
   return false;
 }
 
-template <typename dataType>
 bool GabowTarjan::BFS() {
   std::queue<int> vertexQueue;
 
@@ -69,19 +70,17 @@ bool GabowTarjan::BFS() {
   return Layers[0] != -1;
 }
 
-template <typename dataType>
 void GabowTarjan::HopcroftKarp(unsigned int &matching) {
-  while(BFS<dataType>())
+  while(BFS())
     for(unsigned int vertex = 0; vertex < MaxSize; ++vertex) {
       if(Pair[vertex] == -1) {
-        if(DFS<dataType>(vertex))
+        if(DFS(vertex))
           ++matching;
       }
     }
 }
 
-template <typename dataType>
-dataType GabowTarjan::Distance(dataType maxLevel) {
+double GabowTarjan::Distance(double maxLevel) {
   // Clear the pairing
   Pair.clear();
   Pair.assign(2 * MaxSize, -1);
@@ -159,8 +158,8 @@ dataType GabowTarjan::Distance(dataType maxLevel) {
     this->printMsg("Guessing for " + std::to_string(guessEdge) + "...");
 
     matching = 0;
-    HopcroftKarp<dataType>(matching);
-    // printCurrentMatching<dataType>();
+    HopcroftKarp(matching);
+    // printCurrentMatching();
 
     if(matching >= MaxSize) {
 
@@ -202,7 +201,6 @@ dataType GabowTarjan::Distance(dataType maxLevel) {
   return -1;
 }
 
-template <typename dataType>
 void GabowTarjan::printCurrentMatching() {
   int size = 2 * MaxSize;
   std::vector<int> missedPlaces;
@@ -235,15 +233,14 @@ void GabowTarjan::printCurrentMatching() {
   }
 }
 
-template <typename dataType>
 int GabowTarjan::run(std::vector<matchingTuple> &matchings) {
   // Compute distance.
-  double dist = Distance<dataType>(1);
+  double dist = Distance(1);
   this->printMsg("Computed distance " + std::to_string(dist));
 
   // Fill matchings.
   matchings.clear();
-  auto C = (std::vector<std::vector<dataType>> *)Cptr;
+  auto C = (std::vector<std::vector<double>> *)Cptr;
 
   for(unsigned int i = 0; i < Size1; ++i) {
     if(Pair[i] == -1)
@@ -286,3 +283,5 @@ int GabowTarjan::run(std::vector<matchingTuple> &matchings) {
 
   return 0;
 }
+
+} // namespace ttk
