+++ core/vtk/ttkTrackingFromFields/ttkTrackingFromFields.cpp
@@ -63,7 +63,7 @@ int ttkTrackingFromFields::trackWithPersistenceMatching(
 
   ttk::TrackingFromPersistenceDiagrams tfp{};
   tfp.setThreadNumber(this->threadNumber_);
-  tfp.performMatchings<dataType>(
+  tfp.performMatchings(
     (int)fieldNumber, persistenceDiagrams, outputMatchings,
     algorithm, // Not from paraview, from enclosing tracking plugin
     wasserstein, tolerance, is3D,
@@ -92,21 +92,21 @@ int ttkTrackingFromFields::trackWithPersistenceMatching(
 
   // (+ vertex id)
   std::vector<trackingTuple> trackingsBase;
-  tfp.performTracking<dataType>(
+  tfp.performTracking(
     persistenceDiagrams, outputMatchings, trackingsBase);
 
   std::vector<std::set<int>> trackingTupleToMerged(
     trackingsBase.size(), std::set<int>());
 
   if(DoPostProc) {
-    tfp.performPostProcess<dataType>(persistenceDiagrams, trackingsBase,
+    tfp.performPostProcess(persistenceDiagrams, trackingsBase,
                                      trackingTupleToMerged, PostProcThresh);
   }
 
   bool useGeometricSpacing = UseGeometricSpacing;
 
   // Build mesh.
-  ttkTrackingFromPersistenceDiagrams::buildMesh<dataType>(
+  ttkTrackingFromPersistenceDiagrams::buildMesh(
     trackingsBase, outputMatchings, persistenceDiagrams, useGeometricSpacing,
     spacing, DoPostProc, trackingTupleToMerged, points, persistenceDiagram,
     persistenceScalars, valueScalars, matchingIdScalars, lengthScalars,
